Updated by CBH
David, I've made a number of changes and to the following sections: 
Incentivisation, Culture, Org structure, Org Structure, Culture and Incentives, Automation. Some of mine are line spanners so you'll need to check I've written the JSON correctly.
I think this makes a great start for advice, but we should refine it to be in line with the specific questions; i.e., review through and ensure we've got something covering everything in that section... I don't think we're a million miles away.